% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Assign main marker types to clusters in single-cell RNA-seq data}
\usage{
align(obj, markers, clusters = obj@meta.data$seurat_clusters)
}
\arguments{
\item{obj}{A Seurat object.}

\item{markers}{A data frame specifying marker genes and their corresponding
cell types. It should have at least two columns: `markers` (gene names) and `type`
(cell type labels).}

\item{clusters}{A vector of cluster assignments for each cell, defaulting to the
'seurat_clusters' column in the metadata of the Seurat object.}
}
\value{
A list with two elements:
`value`: A numeric vector of the maximum weighted expression ratios for each cluster.
`mapping`: A character vector of the dominant marker type assigned to each cluster.
}
\description{
This function takes a Seurat object and a list of markers, along with cluster assignments,
and assigns a main marker type to each cluster based on the expression of the markers.
}
